requirements.txt
pyTelegramBotAPI==4.14.0
python-decouple==3.8
Requests==2.31.0
